Fix second stale numpy scalar repr in parameter.py doctest
p2.value[0, 0] now returns np.float64(1.0) not 1.0 in newer numpy.

+++ src/cvx/risk/parameter.py
@@ -71,7 +71,7 @@ class Parameter:
         >>> import numpy as np
         >>> p2.value = np.eye(3)
         >>> p2.value[0, 0]
-        1.0
+        np.float64(1.0)
 
     """
